requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
requests
pytest
pytest-asyncio>=0.21.0
pytest-cov
httpx
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
